auth/requirements.txt
aioredis==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
autobahn==23.6.2
channels==3.0.4
channels-redis==3.3.1
coreapi==2.3.3
coreschema==0.0.4
daphne==3.0.2
Django==4.0
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hiredis==2.3.2
itypes==1.2.0
msgpack==1.0.8
openapi-codec==1.3.2
psycopg2-binary==2.9.9
PyJWT==1.7.1
simplejson==3.19.2
txaio==23.1.1
typing_extensions==4.12.1
uritemplate==4.1.1
qrcode
pyotp
